Update patch set 20
Patch Set 20:

> @Alon: When NeutronPort add comes, there will be no port on switch.
 > Port gets created by Nova after port binding and NeutronPort
 > update. So, attempt to find will always fail, but fine as without
 > viftype no parent ref set at creation. But now, when NeutronPort
 > update comes, we still won't have parentref information. Order will
 > be PortUpdate, TP Create/OFNodeConnectorAdd, updateParentRef. With
 > parentRef information coming much later, most probably last part,
 > aren't we leaving whole lot of activity to be done much later?
 > 
 > Note that am still not fully aware of problem being solved here and
 > some of my questions are coming more from upcoming TrunkPort
 > changes.

@Vishal, not sure I understood the question.
We just don't fill in the parentref if we don't have it yet.
If we do have it (OVS port came before the neutron port), we fill it in and proceed as usual.
As you said, for nova ports the usual flow will be neutron port before termination point.

What activity is left to be done later?

Patch-set: 20