Update patch set 1
Patch Set 1: Code-Review-1

> Build Failed

oh this is FUN - this is the ElanServiceTest again, which started all of this work I'm doing today... feels like going around in circles! ;-)  https://git.opendaylight.org/gerrit/#/c/50824/ fixes this on master of course, and this change has that as its parent, so.. interesting.  I guess / interpret / it looks like that this final clean-up makes a different/older version of Hamcrest appear on the classpath (again) - which only illustrates what a confusing mess this all is, and how important it is to clean all this up once and for all, and put automated non-regression into place.  I'll pause this particular Gerrit for a few days, and come back to it once all the other changes I raised have been merged, then I can make a push for a final clean up to resolve this.

Patch-set: 1
Label: Code-Review=-1