Bug 7669: Add multi-provider network support to NetVirt for L2 Gateway.
Multi-provider network support is needed to be able to support
connecting VLAN based networks to VXLAN based networks,
using L2GW/HWVTEP.

Slides describing the network, and use cases, can be
found here:

https://docs.google.com/presentation/d/13OqyCHwGQABlcno9lVVY2KHVZdNMYz3JiiTRrQj-n98/edit?usp=sharing

Testing
-------
Setup and successfully tested use cases 1-3, as described in
slides above.

Note: There is a corresponding Nova bug that must be fixed
for this to work. Basically, Nova needs to add support
for multi-provider networks. This work is in progress.

Change-Id: Iba4e52210e8c27e0ec35fe68b1fba3970b3a66c8
Signed-off-by: Victor Pickard <vpickard@redhat.com>