Update patch set 1
Patch Set 1: -Code-Review

FYI, this patch does not fix the original issue that I've reported in the bug [1], for which I proposed a patch [2]. I've abandoned the patch [2], since its a networking-odl bug. Ideally, n-odl should not be deleting the networkType attribute prior to network removal.

What are your thoughts on [2]. I feel its good to have it even if we fix the original issue in networking-odl. Any comments?


[1] https://bugs.opendaylight.org/show_bug.cgi?id=7601
[2] https://git.opendaylight.org/gerrit/#/c/50673/2/vpnservice/neutronvpn/neutronvpn-impl/src/main/java/org/opendaylight/netvirt/neutronvpn/NeutronNetworkChangeListener.java

Patch-set: 1