Update patch set 2
Patch Set 2:

One more question - referring to use case 3, is VM1 on a VXLAN network (net1) and VM2 on a VLAN network (net2)?

I'm pretty sure L2GW by itself is working (VM1 pinging a PNF behind the TOR), so since VM2 is SR-IOV and not going through OVS, isn't it the same as a regular PNF on the VLAN network, behind a L2GW?
So our OVS thinks it's a regulard VXLAN network? What am I missing?

Patch-set: 2