Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

@Sridhar

The add() and remove() are common methods.  They will be invoked for any type of network.

If we are not supporting certain network types, they should not be serviced in both add() and remove() methods of NeutronNetworkChangeListener.

Blowing off the provider attributes via an Update() before a Delete() is not a cohesive lifecycle.  I would say the culprit here is the review you have put in the commit message and not ODL.

Are all extensions of a neutron network are blown off with an earlier update() (OR) only provider-type extension of the network is blown off?

Vivek

Patch-set: 3