Update patch set 7
Patch Set 7: Code-Review-1

Tx for having addressed the copy/pasted duplicate code in the *Test classes, that's great progress.  However now we still have a lot of copy/pasted duplicate code between your new FlowEntryObjectsStatefulIPv6 and the existing FlowEntryObjectsStateful *.xtend classes, which I would suggest we try to remove, similarly to what was for the *Test/s.. 

I think in the case of the *.xtend it's already a bit of a mess as-is in master .. ;-) but perhaps we could take this change as the opportunity to clean that up?  The root of the problem is likely that we (Aswin and/or I) had started using "static" methods e.g. in FlowEntryObjectsStateful... in hindsight, that seems like a stupid choice we made, because it makes it impossible to customize all those methods creating the models expected by the tests.  So I would suggest that you remove all the "static" from all the *.xtend (and thus consequently switch the *Test/s to everwhere use e.g. new FlowEntryObjectsStateful().etherFlows() instead of the current FlowEntryObjectsStateful.etherFlows() with "static").  Then you can start to better "re-factor" the *.xtend, and I think avoid A LOT of the current 1700+ lines of duplicates code, by simply putting everything that's different between IPv4 and IPv6 in short methods in FlowEntryObjectsBase, and override them in the FlowEntryObjectsStateful with what they should be for IPv4 and in FlowEntryObjectsStatefulIPv6 for how they're different for IPv6.

NB that in Xtend syntax the equivalent to the Java syntax of "@Override void foo()" is to use "override foo()" instead of "def foo()" (so those Xtend guys made "override" a keyword of the language syntax instead of an annotation like in Java).

Patch-set: 7
Label: Code-Review=-1