Update patch set 5
Patch Set 5: -Code-Review

Removing the -2 - Aswin could you have a look before I share with everyone?
We discussed this internally and it seems to us this proposition will work fine as long as there is a single SG configured on the port.
In cases where there are multiple SGs on the port, we decided to fallback to the current implementation - see Alternatives section for some thoughts on this.

We can do this additional optimization in the future if required, but it seems to complicate the solution so we wanted to avoid it as a first step.

Patch-set: 5
Label: Code-Review=0