Spec to support IPv6 DC to Internet L3VPN connectivity using BGPVPN
This blueprint spec explains architecture for supporting IPv6 Datacenter
to Internet connectivity using BGPVPN in a SDN controlled by OpenDaylight.
Enhance option b) specification as explained in [0] so that it is possible to
have separated vpns for private inter-dc and for internet access. So that
private vpns keep being private, while an other vpn instance will be publicly
opened.

[0] https://drive.google.com/file/d/0BxAspfn9mEi8OEtvVFpsZXo0ZlE/view

Change-Id: I99b6718d6868fe8b0c49081e2f4cf3897ea1e3f6
Signed-off-by: Julien Courtat <julien.courtat@6wind.com>