Bug 7775: Using DJC for NAT Interface-state Listeners
Description : With mdsalutil is having a THREAD_POOL_SIZE=1, all the
add(), update(), and remove() methods handling the application logic were
on the same thread. This in turn was delaying all the interface-state
listeners.

Changes are done to make use of DataJobCordinator so that all the NAT
application logic get executed by a seperate thread so that the main
thread is freed-up for other listener execution.

Change-Id: I0c7f3f70715b01a2dcb54d251a36a2f9ff21357e
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>