Update patch set 3
Patch Set 3:

Setting the port status to ACTIVE only after flows are configured looks good to me.

The opposite direction of delete seems to be missing some use cases, also I imagine you would want to set the port status to DOWN if we receive an ofport staus update or something similar? Do you plan to do this at all or keep it for the future?

Another question - you mentioned this was causing races and failures - is there a way you can verify this fixes the issues you were hitting?

Patch-set: 3