Update patch set 3
Patch Set 3:

(1 comment)

inline... 

> Setting the port status to ACTIVE only after flows are configured
 > looks good to me.
 > 
 > The opposite direction of delete seems to be missing some use
 > cases, also I imagine you would want to set the port status to DOWN
 > if we receive an ofport staus update or something similar? Do you
 > plan to do this at all or keep it for the future?

Right now every port is removed from operational when the neutron port is deleted, here: https://git.opendaylight.org/gerrit/#/c/51554/3/vpnservice/neutronvpn/neutronvpn-impl/src/main/java/org/opendaylight/netvirt/neutronvpn/NeutronPortChangeListener.java

This should eventually clean up every port from operational and in general it makes sense to have a port status around as long as the port exists. However, I am considering modifying this. Here's the issue. The restconf notifications that come over the websocket suck. Big time. Basically, if we subscribe to /neutron/ports then that entire sub-tree is sent every time any port changes status. So, if you have a million ports...you get the idea. Theoretically it would be possible to subscribe for changes multiple times, once for each port that is transitioning from DOWN to ACTIVE, i.e., /neutron/ports/port[uuid=xyz]. However, the way the websocket API is built it requires a TCP connection per subscription which would not be *so* terrible if it weren't for the fact that the websocket library requires a thread per socket (blocking IO). This all sounds, how shall we say, not ideal. I spoke to Robert Varga who said that the subscription/notification model comes straight from netconf and therefor can't be mucked with too much. I'll be researching some of that today but in general I think it's unlikely we'll be able to improve on the sub/notif infrastructure in the near future and not sure it's worth it. So one thing I'm considering is using the OPERATIONAL /neutron/ports only for ports transitioning from DOWN to ACTIVE and once they've transitioned, networking-odl deletes them by rest call. It's a little bit of an abuse of the concepts but no one else is using OPERATIONAL /neutron/ports at this point. Not decided yet on this, still looking around if there is a more elegant and correct fix.

So, regarding the second part of your question, for now I'm only using these states for the DOWN->ACTIVE transition. 

 > 
 > Another question - you mentioned this was causing races and
 > failures - is there a way you can verify this fixes the issues you
 > were hitting?

I do not have a concrete way of testing this at this point. I've asked around. You have one?

Patch-set: 3
Label: Verified=0