Update patch set 3
Patch Set 3:

(1 comment)

> Adding Michael to see if he has some ideas on the code duplication.

You can try to reduce the 292 lines added to FlowEntryObjectsStateful.xtend .. they key to understand here is that those "def" things in the Xtend syntax really are just Java methods (they are not "like" Java methods, they ARE Java methods).  So you may be able to refactor this and put everything that is common in one method, and then call that from another methods that has what's not common.  You can concatenate flows with + like you do in aapWithIpv4AllFlows(), that just creates a Collection.  I don't understand enough to be sure what would make sense to re-use here.. looking at it just visually, it seems like the MatchInfo (whatever that is) in line 605 and 623 and 689 is actually the same (or is it?), so for example one thing that may make sense here is to have another method, something like:

 def choseAGoodNameForWhateverThisIs() { 
     #[
         new InstructionApplyActions(#[
             new ActionNxConntrack(2, 0, 0, 5000, 252 as short)
         ])
     ] 
 }

and then you could just use instructionInfoList = choseAGoodNameForWhateverThisIs().

That's the thing that's so cool about this stuff, it looks very much "declarative", like XML/YAML/JSON, but is actually "imperative" - it really is code, so you can refactor it and call out to helpers to re-use things, even parametrize it (your def methods can have arguments!).

Now to what degree doing the above makes sense, that is at what point is helps re-uses and readability, and when doing too much of this makes it harder to understand what's in a flow from looking at it, to find the right balance for this is something I have to leave up to all of you...

https://git.opendaylight.org/gerrit/#/c/50044/ has a kind of similar discussion, BTW!

 > I would like to keep the UT looking good so we can use it as a
 > template for others.

Totally agree.

Patch-set: 3