Update patch set 7
Patch Set 7:

> It is unclear what the bug is and it attempting to address..
 > 
 > Please offer explanation. i would request if possible to avoid
 > adding more augmentations to the VRFEntry.

Vivek,

Take a look at VrfEntryListener.createFibEntries() it creates the local fib entry and return the local dpn id(s). These ids are later skipped when calling createRemoteFibEntry from within this method

For FIB entries with custom instructions such as FIP and router-gw ip, the local FIB entry is installed using the fib RPC call createFibEntry() and the local dpn+instructions are passed as input to the RPC, separately from the VrfEntryListener code.

In addition, VRF entry is created for each FIP/router-gw ip and when VrfEntryListener detect that this type of entry was added, there is an assumption that VrfEntryListener.createLocalFibEntry() will fail so that LNH with the custom instructions installed earlier will not be overridden with some simple local-nexthop group.

In addition, createLocalFibEntry() does not return any local-dpn so all the dpns in vpn-dpn model are considered for installation of RNH.
As it happens, in VxLAN/GRE external networks there is never a tunnel between DPN and its local ip so the FIB with special instructions is rescued from being overridden by RNH installation. The only outcome is trashing the logs with these errors for the local DPN since no TEP was found 

 2017-02-14 09:04:09,067 | ERROR | nPool-1-worker-0 | ItmManagerRpcService             | 374 - org.opendaylight.genius.itm-impl - 0.2.0.SNAPSHOT | Tunnel not found for source DPN 132696395651039 ans destination IP IpAddress [_ipv4Address=Ipv4Address [_value=10.0.42.52]]
2017-02-14 09:04:09,067 | WARN  | nPool-1-worker-0 | NexthopManager                   | 382 - org.opendaylight.netvirt.fibmanager-impl - 0.4.0.SNAPSHOT | RPC Call to getTunnelInterfaceName returned with Errors []
2017-02-14 09:04:09,067 | TRACE | nPool-1-worker-0 | NexthopManager                   | 382 - org.opendaylight.netvirt.fibmanager-impl - 0.4.0.SNAPSHOT | NextHop pointer for prefixIp 10.64.112.3/32 vpnId 70000 dpnId 132696395651039 is null

For VLAN underlays this magic can't happen so we will end up installing RNH flow also on the local DPN thus overriding the custom instructions.
This seems very wrong to me

Yair has tried to solve it for FIP UC in [1] but Olga has also encountered it while trying to solve another bug in the context of router-gw ip [2] and this bug should go into Boron as well so we must come up with a uniform solution for this.

If you disagree to adding augmentations over vrf-entry please offer an alternate solution for this as it seems very wrong to me to try and guess what is the local dpn-id of some vrf-entry by traversing all kinds of l3/nat models until this address will be found somewhere

[1] https://git.opendaylight.org/gerrit/#/c/51558/
[2] https://git.opendaylight.org/gerrit/#/c/51853/

Patch-set: 7