Update patch set 3
Patch Set 3:

| that is at what point is helps re-uses and readability, 

To re factor this code further I think we will have to use methods accepting IP, mac-address lport tag and build the matches, so that we can match reuse the matches. But I think it will reduce the ease of writing the test case. Now that we can just copy paste the actual to expected , the effort is much less. Now the duplication happens only when new ports are added. For a given port most of the code is not duplicated.

Patch-set: 3