Update patch set 1
Patch Set 1:

Alon, this is the vpp topic branch - a poc to get vpp working within netvirt. This is a topic branch so it doesn't impact master or boron.

The patch is large because it is a sync patch. topic/vpp was pulled from netvirt master 6 weeks ago. The patch covers all the changes since then.

Patch-set: 1