Update patch set 7
Patch Set 7:

The prefix-to-interface is not a new DataStore.  It is already an existing datastore inside VPN and can be reused by NAT Service similar to how NATService re-uses the dpn-to-vpn-list of vpn-instance-op-data.

Also you will need to know the localDPN anyway to ignore programming remote flows on it even by mistake (for any type of prefix) and that is the reason just having a boolean in a VRFEntry may not be sufficient.

Patch-set: 7