Update patch set 1
Patch Set 1:

We still saw a single CSIT run fail [1] for a single VM DHCP in BFD Monitoring. I think this is related to the REST delete that is done, and the BC group isn't being programmed correctly for the ELAN.

We also encountered this NPE in the logs, so guessing this might be the cause.

2017-02-16 11:52:14,376 | ERROR | nPool-1-worker-3 | DataStoreJobCoordinator          | 333 - org.opendaylight.genius.mdsalutil-api - 0.2.0.SNAPSHOT | Exception when executing jobEntry: JobEntry{key='38cd1d79-84c8-4aa1-bc33-5e4d4ded0ace', mainWorker=org.opendaylight.netvirt.elan.internal.ElanInterfaceManager$$Lambda$250/1844113528@3dae6914, rollbackWorker=null, retryCount=0, futures=null}
java.lang.NullPointerException
        at org.opendaylight.netvirt.elan.internal.ElanInterfaceManager.setupStandardElanBroadcastGroups(ElanInterfaceManager.java:1047)


[1] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-carbon/96/archives/log.html.gz#s1-s4-s2-k1-k4-k28-k2-k1-k3-k1-k17-k4

Patch-set: 1