Update patch set 3
Patch Set 3:

In theory we can develop dedicated protocol by odl neutron northbound.
Essentially it would be a variant of websocket notification.

But I wouldn't like to go for developping dedicated protocol because this notification issue is not only for port status, but also for other neutron resource and hostconfig. Until netconf websocket notification is ready for us, we can use polling as work around. polling is much easier than dedicated protocol.

Patch-set: 3