Update patch set 2
Patch Set 2:

> In yesterday's meeting, we had a discussion,since we are
 > deprecating even the async listener bases, should we use that..
 > Even in my parent patch, I have used AsynListenerBase only..would
 > like to get a confirmation from someone, whether this is enough to
 > close the clean up activity

That is a separate activity and we're yet to decide if going to let go abstract altogether or create a newer better AsyncBase abstraction. That can come in separate patch, depending on decision. This one is priority because DCL is already deprecated and will be removed soon.

Patch-set: 2