Bug 7815: Using DJC for VpnManager Interface-state Listeners
With mdsalutil is having a THREAD_POOL_SIZE=1, all the add(), update(),
and remove() methods handling the application logic were on the same
thread. This in turn was delaying all the interface-state listeners.

Changes are done to make use of DataJobCordinator so that all the
VpnManager application logic get executed by a seperate thread so
that the main thread is freed-up for other listener execution.

Change-Id: I926e0742fdc7955450c9423ce5026f9177ce21bb
Signed-off-by: ehvkand <hanamantagoud.v.kandagal@ericsson.com>