Update patch set 26
Patch Set 26: -Code-Review

As we discussed, we have to maintain consistency between what is inside the VRFTable and what is published to BGP, for both L3 and L2 bgpvpns.

The movement of gateway_mac_address itself can be in a different subsequent review but it has to be eventually moved as conveyed above.

Patch-set: 26
Label: Code-Review=0