Ethernet VPN (EVPN_RT5) Enhancement to Non-NAPT to NAPT
This review (1st Review) is part of EVPN_RT5 support in NAT feature.

Changes are made in the NAT code for accomplishing the below

Intra DC Communication:
=========================
For Non-NAPT to NAPT Traffic, Tunnel id will be setting with
Router lPortTag which will be carved out per router if it is part of
L3VpnOverVxlan else tunnel-id will be setting with existing router-id.

Change-Id: Idc5e8b1956311541ae7bae255d023c5f3543f9a3
Signed-off-by: karthikeyan <karthikeyan.k@altencalsoftlabs.com>