Update patch set 4
Patch Set 4:

Hi Miguel,

We were discussing about ODL allocating a pseudo-port per subnet of the router, and internally using that to retain the remote flows required on a designated switch called NAPT Switch that represents a router that has an external gateway connection.

The cloud-service-chain uses pseudoports for intervpnlinks in a similar way.

Vivek

Patch-set: 4