Update patch set 27
Patch Set 27:

> My only concern with this is removal of old config and yang files.
 > This is already present in Boron release and now being removed
 > without being deprecated for a release first. Any deployments that
 > changed the default [disabled autoconfig] would be impacted, more
 > so when upgrading. Practice is to deprecate older, have it co-exist
 > with new one and then remove a release later. If not, we need to
 > highlight this in documentation and release notes for Carbon.

I don't see an easy solution for keeping the old config file along with the new one, and having just the old one is a bit strange (using config from one bundle to configure a different bundle). I would prefer to highlight this change in the docs instead so people are aware of it. Could you point me to where I should write about this change? Thanks.

Patch-set: 27