Update patch set 6
Patch Set 6:

Hi Olga,

1) The add method of TunnelInterfaceStateListener doesn't handle the fib entries in case the tunnel is not UP yet. As well, the tunnel state updating (from Down to UP) is not covered at all.

[Vivek] The Down to Up case is covered in the update() method inside TunelInterfaceStateListener.
        boolean isTunnelUp = (tunOpStatus == TunnelOperStatus.Up);
        handleTunnelEventForDPN(update,
            isTunnelUp ? UpdateRouteAction.ADVERTISE_ROUTE : UpdateRouteAction.WITHDRAW_ROUTE,
            TunnelAction.TUNNEL_EP_UPDATE);




2) To update the FIB table, TunnelInterfaceStateListener looked through different DS models, for example odl-interface-meta:dpn-to-interface-list, l3vpn:vpn-instance, l3vpn:vpn-interfaces . Is there a guarantee that all the operational models are ready before this tunnel listener gets the events? 

[Vivek]  

The state-tunnel-list listened in the above file is created only when both ITM and InterfaceManager operational datastores for the tunnel-interface is up.  And the above file listens on the same.

So the structure odl-interface-meta:dpn-to-interfaces-list will be available.  But there is no guarantee vpn-instance / vpn-interface will be available.  If they are not available, but come later that is already handled.  As per this bug, vpn-instance and vpn-interface are available, but the tunnel info is not yet available.  This case is one of series of tunnel use-cases already handled inside the above file.  

So the code there should have already worked for us, if not we just need to debug that file i.e, TunnelInterfaceStateListener.java. 

Further if you notice, the above file safely handles the management of operation-vpn-interfaces, updation of fib entries..

Vivek

Patch-set: 6