Update patch set 6
Patch Set 6:

Vivek,
No we do not support dynamic update of NAT mode now, we do require a restart of the controller. It think it will not possible right now to maintain the existing translations. Probably we can take this up when the new implementation is more stable and when we have a specific requirement.

Patch-set: 6