Update patch set 6
Patch Set 6:

Faseela,

The existing translation will be lost. The new compute will have the appropriate rules (if this is the first port in the compute node for a router subnet related rules will be added) for the vm to reach the NAPT switch or do the translation if the node is  a NAPT switch. Connections like TCP will timeout and needs to reconnect.

Patch-set: 6
Label: Verified=0