Update patch set 28
Patch Set 28:

> Hi Yakir,
 > 
 > It will be good to have a minimal spec for this one to understand
 > the overall implications.
 > 
 > Vivek

Vivek,

I'm not sure there is enough design changes in this patch to justify a spec.
I'll try to summarize the changes so the content of this patch would be more clear:

The current auto-tunnel implementation in netvirt creates a new transport zone for each private VxLAN network and for each router if there is at least one VxLAN private network connected to the router. Transport zone and VTEP creation is optimized to include only OVS nodes that have at least one VM spawned in the VxLAN network/connected to the router.

The change introduced in this patch is mainly to support SD-WAN (unimgr) use cases. The main difference is that the trigger to create/update the transport-zone with new TEP is based on common netvirt model updates e.g. ELAN interface, VPN interface rather than relaying on the neutron port model.

In addition, this patch introduces several non-functional optimization such as:
 * minimize the amount of DS access by avoiding read/write operations of subtrees
 * minimize the amount of auto-tunnel listeners to vpn-dpn and elan-dpn models
 * avoid races between tunnel creation and ELAN BC group update with tunnel buckets (the race exist in the current implementation)

Patch-set: 28
Label: Verified=0