Update patch set 6
Patch Set 6:

Hi Olga,

We went through these use-cases with StateTunnelList.  After trials over it we realized that it is better to program the flows on TunnelAdd event instead of waiting until a Tunnel UP.

Furhter in a scaled cloud environemnt if a tunnel keeps flapping, this will unnecessary trigger a host of flow removals and flow adds and can potentially leave the OVS Datapath in an indeterministic state.

Updatee events with StateTunnelList is very sporadic.

So please consider doing this logic inside Tunnel_EP_ADD itself.

Vivek

Patch-set: 6