Update patch set 4
Patch Set 4:

> This patch will have impact on timing. This doesn't mean this patch
 > is introducing a bug, but uncovering ones already there, probably
 > few fixed by Guy's patch [https://git.opendaylight.org/gerrit/#/c/51913/].
 > Without that patch, all the AsyncDataTreeListeners will end up
 > using common thread and any listener that takes too long will be
 > delaying others.
 > 
 > Assuming this patch is unconvering these issues, we have to make a
 > call if we want to get this in and then fix issues or block this
 > till all issues are resolved.

Vivek is working on the DHCP bugs from the list I mentioned.
Peri mentioned [1] might also help with those, let's see how it goes after that is merged.

If this patch fails consistently, I prefer using it unmerged as a way to resolve the DHCP issues. If it only fails sporadically, we can probably merge and continue following up on the DHCP races.

[1] https://git.opendaylight.org/gerrit/#/c/52277/

Patch-set: 4