Bug 7863 - Add Layer 4 Match for flow entries for TCP/UDP security group rule with no min/max
The methods Ingress/EgressAclService::ingress/egressAclTcp/Udp
incorrectly assume that portRangMin/Max values are always specifed.
This is not true when using openstack command lines to
define SG rule for TCP/UDP with no min/max ports specified.
As a result, no layer 4 match are added to the flow entries.

The fix is to add logic to handle null min/max values for the
above scenarios.

Change-Id: I537f51fdd56d8aca3711a38cd5fd2c5eba5d9af9
Signed-off-by: Vinh Nguyen <vinh.nguyen@hcl.com>