Bug 7831 : BgpRouter receives unnecessary events
Suppressed unwanted advertisement of routes, that were being advertised
due to a
breakage in the VPN Engine.
When BGP Client is not configured, route/vrf advertisement used to fail
and throw lot of exception. If BGP Client is not connected, the
process of dispatching routes/vrf to DC-GW via BGP Client should not
happen
at all.

Change-Id: I3d85c8bca6bf3462d8a7694a8f585bf437aea11c
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>