Update patch set 12
Patch Set 12:

(9 comments)

Replied to comments.
I still have an open issue, on how to distinguish between BGPVPN and internal VPN usecases when creating a subnet in the external network.
Today, the decision of the VPN type based on VpnID != RouterID , is no longer valid, because in external subnet case, the VpnID is indeed different than the RouterID.

Patch-set: 12