Bug 7714: VPN Operational Interfaces not getting removed at all.
This can be a potential problem for temporary CSIT failures
that happens on VPNService suite.

The VPNInterface is not getting removed at all permanently from
VpnInterface Op DS. When VM interface has secondary adj like extra route
or invisible ip , the order in which cleanupvpninterface worker DJC is
called not guaranteed.

The regression was caused by review:

https://git.opendaylight.org/gerrit/#/c/51224/
VPNInterfaces not deleted from Oper DS

Change-Id: I1b2a1d139030f707cec93eb5a5e6a690e010e336
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>
Signed-off-by: ehvkand <hanamantagoud.v.kandagal@ericsson.com>