Update patch set 3
Patch Set 3:

> This also needs features.xml changes. Though we'll probably need
 > waiver for it as karaf feature freeze is M3.

Vishal, do you think we can merge this commit as is, before M4?

In parallel Tali can raise a feature.xml review and we will ask for a waiver on it.

After that, we can get the entire spec reviewed and the full code merged (by M5).

Patch-set: 3
Label: Verified=0