Enable SubnetRoute in non BGPVPN usecase
First commit in a series of commits, part of the discovery
of directly connented PNFs in Flat/Vlan provider networks.

This commit enables the creation of VrfEntry with a subnetroute
augmentation, in non BGPVPN use case, with respect to the recently
changes made by the hairpinning spec.

A VrfEntry is created in both external networks and private tenant
networks.

Change-Id: Ida689b6e5317e45e6a9c8a4f6887e5e9f2438161
Signed-off-by: Tomer Pearl <tomer.pearl@hpe.com>

<WIP> Discovery of directly connected PNFs.
First commit this this feature

TODO: add list of changes

Change-Id: Ida689b6e5317e45e6a9c8a4f6887e5e9f2438161
Signed-off-by: Tomer Pearl <tomer.pearl@hpe.com>