Update patch set 12
Patch Set 12: -Code-Review

I don't want to block this spec for M4 as it bring useful feature..

At the same time, the yange change that Tomer himself has mentioned is not sufficient to address this case.. So what is the final take?

Removing my vote to unblock!

Patch-set: 12
Label: Code-Review=0