Update patch set 8
Patch Set 8:

> (1 comment)

In response to Juanma, I understand the concerns but I think that is an SFC discussion and not Netvirt one. I think that the equivalent issue for the netvirt classifier is covered in the second to last paragraph of the Integration with Genius section of this blueprint.

Logical SFF in SFC project currently assumes a 0-Day full mesh config and will need to be impacted for dynamic tunnel handling. Logical SFF is aware if two SF in a sequence are on the same node and handles the situation as expected.

Patch-set: 8
Reviewer: Gerrit User 6230 <6230@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0