Update patch set 8
Patch Set 8:

It may be worth to explicitly mention the lack of need for an outbound classifier compared the old netvirt classifier. In a way, the fact that there is no outbound classifier requires an specific configuration of other services so that packets can be forwarded at chain exit. Moreover, with the implementation of flow based tunnels it maybe possible to also have an outbound classifier following the same approach that of the old netvirt classifier. What do you think?

Patch-set: 8