Bug 7931:  SubnetRoue re-election to be triggered on disconnected nodes
Subnet Route reelection algorithm is triggered today only when
the anchor ports on the subnet go to DOWN state (or) if such
anchor ports on the subnet are being removed.

However, there are situations where a switch hosting the anchor
port can be disconnected on the openflow control-path and can
continue to remain so indefinitely.

Under such circumstance, new non-neutron IPs become undiscoverable
as the switch hosting the anchor port cannot punt discoverable IP
packets to the controller.

So this fix attempts to always trigger reelection of nodes for
subnetRoute hosting as soon as the current anchor node(or switch)
is detected to be disconnected from the controller.

Change-Id: Idd8be8e2f1940c4840d7d52726ac5187ed33d28f
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>
(cherry picked from commit f9d156dc32aff937425eb586bee3ceefda069091)