Update patch set 3
Patch Set 3:

We did want qosservice to support non-neutron ports in general - but as the code already assumes support just for neutron ports, this patch is fine just for exception handling.

I assume there is a parallel effort to generalize the qosservice to not have any neutron port dependency? Unimgr would probably need this..

Patch-set: 3