Bug 7984: IDLE_TIMEOUT check required in onFlowRemoved.
Description: When FlowRemoved notification is received,
We are taking care of deleting the INBOUND_NAPT_TABLE entry
and releasing the allocated external port for an external-ip.
This is valid only if onFlowRemoved received due to OUTBOUND_NAPT_TABLE
removed beause of IDLE_TIMEOUT. When Switch goes down, we re-elect
a new NAPT switch and re-install ongoing session flows. It is required
not to release the allocated external port during Switch down, but
should be done only during IDLE_TIMEOUT.

Initially, we have the check but been removed with review
https://git.opendaylight.org/gerrit/#/c/49214/. Also plugin support
to get the RemovedReason was not available with onFlowRemoved.

Now, we have the plugin support to get the removed reason which can be
used. https://git.opendaylight.org/gerrit/#/c/50335/

Change-Id: I5708fbd1755719ea0cdabe29eaf0787076a46f24
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>