Update patch set 7
Patch Set 7:

Thanks Michael, I think this looks good, but I'm not sure qosservice is the right place for this utility.

As mentioned, this behavior is done in multiple places across netvirt - we should correct them all to use the same - I was thinking to merge the previous patchset as a bug fix, and do this refactoring across all modules in a separate commit.

Vivek/Vishal what do you think?

Patch-set: 7
Label: Verified=0