Bug 7988 - Datapath not configured after cluster r
Elan Instance , Physical Switch , L2gateWay and L2gatewayConnection all
have dependency among each other.But if upon cluster reboot the order is
not proper (L2gw cache is updated at the last , then flows are not being programmed as expected.
To take care of the situation
Introduced new ElanInstance Listner (Was already present but not
registered with xml to be instantiated )
Handle scenario if L2gW shows up last after cluster reboot .
Handle scenario if l2gwConnection is updated but Elan instance is not
available.

Change-Id: I6194d7cb22d6659803b17078d57c8ad694c0c201
Signed-off-by: Akash Kumar Sahu <a.k.sahu@ericsson.com>