Update patch set 2
Patch Set 2:

Marcos, Miguel - can you have a look at the cloudservicechain change?

This should be a bug, but it seems VPNServiceChainHandlerTest#testprogramScfToVpnPipeline() expects 2 flows to be installed - this changes the behavior so only 1 flow is installed.

Since I'm not familiar with this use case I don't want to just fix the unitest as I assume it was correct to expect 2 flows?

Patch-set: 2