Update patch set 12
Patch Set 12:

> I will need some time to go through this one..
 > 
 > If federation plugin iself is going to be redesigned in Nitrogen,
 > won't this federated l2/l3 connectivity be best done in Nitrogen
 > timeframe.
 > 
 > However, I donot want to block this spec from merging.

Hi Vivek,

Even though the implementation might be redesigned going forward, this feature is working and has APIs/dependencies in Boron.
It is also the first plugin of the federation upstream project, and is required to merge to allow testing such as CSIT for federation (required for release), and to minimize potential breakage.

We kindly request to merge this spec and its features this week so we can move forward, if you prefer to postpone the review itself that's ok too (it seems no one outside of HPE has cycles to review this).
https://git.opendaylight.org/gerrit/#/c/52178/
https://git.opendaylight.org/gerrit/#/c/52860/

Patch-set: 12
Reviewer: Gerrit User 670 <670@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0